feat: wire SequentialAgent, replace chat.py orchestration (re-me1f)
Create backend/pipeline.py with ChatPipeline class that:
- Wires context, reasoning, and response ADK LlmAgents into a SequentialAgent
- Centralizes the orchestration logic (context gathering, refinement loop,
  external data fetching, reasoning, response generation)
- Provides run() for non-streaming and run_stream() for SSE streaming
- Emits PipelineEvent objects for stage transitions and token streaming

Simplify backend/routers/chat.py:
- Remove ~400 lines of inline orchestration from chat() and chat_stream()
- Both endpoints now delegate to ChatPipeline
- Extract _build_pipeline(), _fetch_history(), _persist_exchange() helpers
- Keep endpoint definitions, history management, DB persistence, SSE formatting

Move data-gathering helpers to pipeline module:
- _fetch_relevant_things, _fetch_with_family, _fetch_user_relationships
- _fetch_user_thing, _fetch_gmail_context, _sql_things_count

Update all test patch paths from backend.routers.chat to backend.pipeline.

Part of #91
